Add boxes, still some bugs

+++ flowmanager/Workflow.py
@@ -6,6 +6,7 @@ import subprocess
 import psutil
 
 from WorkflowProfiler import profile
+import resource_monitor
 
 class Workflow():
         # note that the workflow should accept a single input gzipped tar file
@@ -16,21 +17,22 @@ class Workflow():
             self.makeflow_dir = os.path.dirname(self.makeflow)
             self.expected_input_name = expected_input_name
 
-        def run(self, input_file, output_directory, error_dir, proc_stats):
-            proc = multiprocessing.Process(target=self.__run, args=(input_file, output_directory, error_dir, proc_stats))
+        def run(self, input_file, output_directory, error_dir, proc_stats, resources):
+            proc = multiprocessing.Process(target=self.__run, args=(input_file, output_directory, error_dir, proc_stats, resources))
             proc.start()
             return proc
 
         # takes the input file, runs it with the given workflow
         # puts the output file in output_directory
-        def __run(self, input_file, output_directory, error_dir, proc_stats):
+        def __run(self, input_file, output_directory, error_dir, proc_stats, resources):
             # parse the filename
             (inputname, filehash, ext) = parse_filename(os.path.basename(input_file))
             workflow_directory_name = os.path.abspath(f"makeflow-{inputname}-{filehash}")
 
             # copy the makeflow code and input file
-            shutil.copytree(self.makeflow_dir, workflow_directory_name)
-            shutil.copyfile(input_file, os.path.join(workflow_directory_name, self.expected_input_name))
+            if not os.path.isdir(workflow_directory_name):
+                shutil.copytree(self.makeflow_dir, workflow_directory_name)
+                shutil.copyfile(input_file, os.path.join(workflow_directory_name, self.expected_input_name))
 
             # change to the new makeflow directory
             prev_dir = os.getcwd()
@@ -40,26 +42,33 @@ class Workflow():
             prc = subprocess.Popen(["makeflow", os.path.basename(self.makeflow)], stdout=subprocess.DEVNULL)
 
             # monitor memory and cpu usage
-            cpu_usage, mem_usage, ccpu, cmem = profile(prc.pid, workflow_directory_name, interval=0.5)
+            cpu_usage, mem_usage, ccpu, cmem, reason = profile(prc.pid, workflow_directory_name, interval=0.5, resources=resources)
 
             # wait for it to finish to get return code
+            for p in psutil.Process(prc.pid).children(recursive=True):
+                p.wait()
             prc.wait()
 
-            # create output filename
-            output_filename = inputname + "-" +  filehash + ".tar.gz"
+            if not reason:
+                # create output filename
+                output_filename = inputname + "-" +  filehash + ".tar.gz"
 
-            if prc.returncode:
-                shutil.move("output.tar.gz", os.path.join(error_dir, output_filename))
+                if prc.returncode:
+                    shutil.move("output.tar.gz", os.path.join(error_dir, output_filename))
+                else:
+                    shutil.move("output.tar.gz", os.path.join(output_directory, output_filename))
+
+                os.chdir(prev_dir)
+                shutil.rmtree(workflow_directory_name)
+
+                input_dir = os.path.dirname(input_file)
+                os.rename(input_file, os.path.join(input_dir, inputname + "-post-" + filehash + ext))
             else:
-                shutil.move("output.tar.gz", os.path.join(output_directory, output_filename))
+                os.chdir(prev_dir)
+                input_dir = os.path.dirname(input_file)
 
-            os.chdir(prev_dir)
-            shutil.rmtree(workflow_directory_name)
 
-            input_dir = os.path.dirname(input_file)
-            os.rename(input_file, os.path.join(input_dir, inputname + "-post-" + filehash + ext))
-
-            stats = {"pid": os.getpid(), "exitcode": prc.returncode, "memusage": mem_usage, "cpuusage": cpu_usage, "cluster_cpu": ccpu, "cluster_mem": cmem}
-            proc_stats.put(stats)
+            stats = {"pid": os.getpid(), "exitcode": prc.returncode, "memusage": mem_usage, "cpuusage": cpu_usage, "cluster_cpu": ccpu, "cluster_mem": cmem, "reason": reason}
+            proc_stats.put((resources, stats))
